Separate multi-nubbin image into individual nubbins
The purpose of this script is to separate images that contain several nubbins that were photographed together into equally sized images that each contain one nubbin. This is achieved using cv2. 

Images containing up to six nubbins are reduced in size and separated into approximately equally sized slices of which each contained one nubbin. This is performed using a Python script ‘SaveSeparatedNubbins.ddmmyy.py’. For each day that nubbin images had been acquired a separate script is used. 

The script uses the python libraries
• cv2 for the image handling 
• TkInter to manage the event-driven user interface and
• PIL as this interfaces well with TkInter.
• pandas is only used because it allows for convenient file handling

The script uses global variables throughout – this is not pretty and makes it less scaleable. However, this version with the global variables works.

Full sized jpg images are loaded using the cv2.imread command and converted into RGB (red-green-blue) format from BGR (the colour order that is used in OpenCV for historical reasons). Each image is resized using cv2.resize to a size of 1600x1200.

A cv2 canvas is constructed with four buttons at the bottom. The buttons [Increase Shift] and [Decrease Shift] are used to shift a grid of seven yellow lines to the left or right until the left yellow line is just to the left of the leftmost bit of the leftmost nubbin or mounting place). Once the seven lines are in the right position, the distance between the lines can be increased using the button [Increase Width] or decreased using the button [Decrease Width]. 

When the seven lines separating the six nubbin are well positioned), the [Save] button can be pressed, the nubbin slice images are saved, the window is closed and a new image is loaded. A log file is generated with information about the saved nubbin slice.